requirements.txt
alembic==1.2.1
cffi==1.13.1
Click==7.0
cryptography==2.8
Flask==1.1.1
Flask-Markdown==0.3
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
Pillow==6.2.1
pycparser==2.19
PyMySQL==0.9.3
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
python-slugify==4.0.0
six==1.12.0
SQLAlchemy==1.3.8
text-unidecode==1.3
Unidecode==1.1.1
Werkzeug==0.16.0
WTForms==2.2.1